Disable doclint when building under JDK8.
Because it complains about everything and fails the build.

Also fix something that my IDE was marking as an error: the <overWrite> element from the configuration for the maven-dependency-plugin isn't valid for the unpack-dependencies goal (there's <overWriteIfNewer>, but it's true by default), so I removed it.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=82468188