Use J2ObjC's @RetainedWith annotation instead of @Weak on ImmutableMap.values to prevent crashes on iOS.
TESTED=Created a map inside an autorelease pool and saved a reference to the
values collection. Checked that the values collection functions correctly
outside the autorelease pool. Checked that both the values collection and the
map are deallocated by the autorelease pool when not holding a strong reference
to the values collection. Did this test with ImmutableMap.of(<2 entries>) and
ImmutableMap.copyOf(<EnumMap with 2 entries>).

RELNOTES=J2ObjC: Fixes crashes from use of `ImmutableMap.values()`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=235206171