Remove `@Beta` from `EventBus` and related classes.
Note that we're primarily removing this because we do not intend to make further changes to the package, and we're discouraging its use. (See https://github.com/google/guava/commit/82a3ecf42cf63f2ea4db5c605d80c4478d8877ae.) We're not going to actually `@Deprecate` it, though, because that would be unnecessarily disruptive especially given that migrating to another solution is not necessarily straightforward.

Fixes https://github.com/google/guava/issues/3360.

RELNOTES=n/a
PiperOrigin-RevId: 369293158