Use `long` multiplication `FarmHashFingerprint64`.
This doesn't change behavior: The modified methods run only on small arrays, so there is no danger of overflow.

Still, in general, it's probably bad style to perform `int` math and then autoconvert the result to a `long` (which I _think_ is what we're doing here?), and I think there may even be an Error Prone check for a similar case(??). So let's not.

(part of https://github.com/google/guava/pull/5989)

RELNOTES=n/a
PiperOrigin-RevId: 438926612