Declare that `Converter.apply` accepts `A` and returns `B`, as its `Function<A, B>` supertype indicates.
Until now, we'd gone with `@CheckForNull A` and `@CheckForNull B`, probably motivated by some combination of:

- the feeling that those are safer from a Kotlin perspective (since the `@CheckForNull` annotation ensures that Kotlin won't require a runtime null check on the value passed to `apply`)
- the fact that `NullPointerTester` would want a method that accepts plain `A` to throw `NullPointerException` when `null` is passed, which `Converter.apply` does not

The former is only a minor concern, since almost no one calls `Converter.apply` (as opposed to `Converter.convert`). The latter we can address in `NullPointerTester` itself, and this CL does so.

The upside is that we're not longer "cheating" in how we declare the method. The new signature matches what we'd need for a Kotlin version of the class.

(Also, this CL removes a `@CheckForNull` annotation and suppression in the class defined by `convertAll`. That annotation (or at least suppression or cast) was probably necessary at one point, but it's not necessary now that our checker understands that `convert(nonNullValue)` always returns a non-null value, even though `convert` can return null values in general (aka "`@PolyNull` behavior"). That checker behavior is also the reason that we don't need an annotation (or suppression or cast) even on our new declaration of `apply`.)

RELNOTES=n/a
PiperOrigin-RevId: 512608753