Refactor - Make Anonymous inner classes lambda
This copy of #6424 omits the change to `CacheBuilder`, as discussed in an existing comment in that file. It also omits the change to `LongAddables`, partially to keep that file closer to upstream and partially because a comment in that file discusses classloading, which might (for all I know) differ between anonymous classes and lambdas.

Fixes #6424

RELNOTES=n/a
PiperOrigin-RevId: 524835176