Update mainline `ImmutableSortedSet.construct` docs to contain the same warning as the backport.
I think the `k < contents.length` case became just as dangerous the `k == contents.length` case way back (maybe even from the beginning; I haven't looked), but we never updated this comment. (It was temporarily made more clearly dangerous (and documented) along with the backport copy in cl/150805495, but the mainline portion of that CL was subsequently rolled back.)

(The main danger would occur if a random caller were to mutate the array later. I could also imagine more subtle problems, like if our `Builder` were to try to reuse the array by appending to it: That would be bad because previously built sets could now retain references to the newly added items. It would also be fragile because that builder would still need to make a copy of the array before sorting it, since the sort might shift a new element over into one of the slots that stores the original set's elements. And there could still be other problems, or I could still be confused about all this....)

(This came up in cl/542316636, but I'm splitting it off.)

RELNOTES=n/a
PiperOrigin-RevId: 548254479