Make `Futures.getChecked` prefer constructors with a `Throwable` parameter.
It still prefers constructors with a `String` parameter over anything else, but to break ties among the constructors that meet that criterion, it now looks for a `Throwable` parameter, too.

This helps with exception classes that accept a `Throwable` and do something with it beyond setting it as the `cause`.

RELNOTES=`util.concurrent`: `Futures.getChecked`, which continues to prefer to call constructors with a `String` parameter, now breaks ties based on whether the constructor has a `Throwable` parameter. Beyond that, the choice of constructor remains undefined. (For this and other reasons, we discourage the use of `getChecked`.)
PiperOrigin-RevId: 563089337