Use `assertThrows` in a few more places.
(followup to cl/563137030)

This CL (if I did it right) covers `@GwtCompatible` tests whose `assertThrows` usages all fall within `@GwtIncompatible` sections. That way, I avoid mixing `assertThrows` and `try`-`fail`-`catch` within the same file. Not that we couldn't do that, but that it seemed relatively more likely to confuse people (and also slightly harder for me to prepare a CL for :)).

RELNOTES=n/a
PiperOrigin-RevId: 570474348