Withdraw `SequencedCollection` overrides from the public copy of `ImmutableSortedSet`.
As far as external users are concerned, this CL is a rollback of cl/575459262.

The problem that cl/575459262 was addressing looks like it may be a quirk of Google's internal build setup. We may end up wanting to remove these overrides entirely, so we don't want to commit to them externally by publishing a release with them. Or we may decide to include them both internally and externally. For now, we'll be conservative.

RELNOTES=n/a
PiperOrigin-RevId: 582099306