Change the return types of `transitiveClosure()` and `reachableNodes()` to `Immutable*` types.
One of those changes is only a change in the declared return type; the other is a behavior change:
- `reachableNodes()` already returned an immutable object, even though that was not reflected in the declared return type.
- `transitiveClosure()` used to return a mutable object.

RELNOTES=`graph`: Changed the return types of `transitiveClosure()` and `reachableNodes()` to `Immutable*` types. `reachableNodes()` already returned an immutable object (even though that was not reflected in the declared return type); `transitiveClosure()` used to return a mutable object.
PiperOrigin-RevId: 607051970